requirements.txt
streamlit
pandas
altair
crewai
langchain_groq
crewai-tools
python-dotenv
litellm
